creating the activities directory for a better organization
