--- app/src/main/java/com/example/task_app/CreateAccount.java
@@ -1,30 +1,0 @@
-package com.example.task_app;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.example.task_app.databinding.ActivityCreateAccountBinding;
-
-public class CreateAccount extends AppCompatActivity {
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_create_account);
-
-        final Button button = findViewById(R.id.Create_account_2);
-
-        button.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                // Code here executes on main thread after user presses button
-                Intent intent = new Intent(v.getContext(), SyncUp.class);
-                startActivity(intent);
-            }
-        });
-    }
-}
