+++ app/src/main/java/com/example/task_app/MainActivity.java
@@ -5,6 +5,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
 
+import com.example.task_app.activities.LogIn;
+
 public class MainActivity extends AppCompatActivity {
 
     @Override
