--- app/src/main/java/com/example/task_app/Pantalla_carga.java
@@ -1,29 +1,0 @@
-package com.example.task_app;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-public class Pantalla_carga extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_pantalla_carga);
-
-        int Tiempo = 1000;
-
-        /**
-         * Pantalla de carga de la aplicacion
-         * */
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                startActivity(new Intent(Pantalla_carga.this, MainActivity.class));
-                finish();
-            }
-        },Tiempo);
-    }
-}
