--- app/src/main/java/com/example/task_app/SyncUp.java
@@ -1,27 +1,0 @@
-package com.example.task_app;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-public class SyncUp extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sync_up);
-
-        int Tiempo = 1000;
-
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                startActivity(new Intent(SyncUp.this, LogIn.class));
-                finish();
-            }
-        },Tiempo);
-
-    }
-}
