+++ app/src/main/java/com/example/task_app/activities/HomeApp.java
@@ -1,0 +1,16 @@
+package com.example.task_app.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.task_app.R;
+
+public class HomeApp extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_home_app);
+    }
+}
