+++ app/src/main/res/layout/activity_pantalla_carga.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Pantalla_carga">
+    tools:context=".activities.Pantalla_carga">
 
     <TextView
         android:id="@+id/textView7"
